requirements.txt
openai>=1.30.0
requests>=2.31.0
Pillow>=10.0.0
huggingface_hub>=0.23.0
